Some var changes for D motor and line lengths
Check settings on motor spool direction and then change these settings above. I think D may be flipped in the on line 352 but should not be. By changing that back and altering
he mech setting it should fix the parabolic motion.

also check line 126 which is used to set the steps per spool radian... this may be different for us - also check microstep. Line 119 sets the line length which may need to be adjsted